Codechange: move Survey-related functions to their own file
In a later commit, this will be shared by the crashlog, as the
crashlog too needs to survey the system. This makes it easier in
the next commit to see what actually changes, and what is just
a plain old move of code.
