+++ src/os/unix/survey_unix.cpp
@@ -8,16 +8,13 @@
 /** @file survey_unix.cpp Unix implementation of OS-specific survey information. */
 
 #include "../../stdafx.h"
-
-#include "../../3rdparty/nlohmann/json.hpp"
+#include "../../survey.h"
 
 #include <sys/utsname.h>
 #include <thread>
 #include <unistd.h>
 
 #include "../../safeguards.h"
-
-extern std::string SurveyMemoryToText(uint64_t memory);
 
 void SurveyOS(nlohmann::json &json)
 {
